feat: expreimental desktop mirror
