+++ airsync-mac/airsync_macApp.swift
@@ -91,6 +91,9 @@ struct airsync_macApp: App {
                         if !appState.isNativeMirroring {
                             dismissWindow(id: "nativeMirror")
                         }
+                        if !appState.isNativeDesktopMirroring {
+                            dismissWindow(id: "nativeDesktopMirror")
+                        }
                         if appState.activeCall == nil || appState.callNotificationMode != .popup {
                             dismissWindow(id: "callWindow")
                         }
@@ -128,6 +131,13 @@ struct airsync_macApp: App {
                 openWindow(id: "nativeMirror")
             } else {
                 dismissWindow(id: "nativeMirror")
+            }
+        }
+        .onChange(of: appState.isNativeDesktopMirroring) { oldValue, newValue in
+            if newValue {
+                openWindow(id: "nativeDesktopMirror")
+            } else {
+                dismissWindow(id: "nativeDesktopMirror")
             }
         }
         .commands {
@@ -252,6 +262,21 @@ struct airsync_macApp: App {
         }
         .windowResizability(.contentSize)
         .defaultSize(width: 320, height: 680)
+        .windowStyle(.hiddenTitleBar)
+        .defaultPosition(.center)
+
+        Window("Desktop Mirror", id: "nativeDesktopMirror") {
+            if #available(macOS 15.0, *) {
+                NativeDesktopMirrorView()
+                    .environmentObject(appState)
+                    .containerBackground(.ultraThinMaterial, for: .window)
+            } else {
+                NativeDesktopMirrorView()
+                    .environmentObject(appState)
+            }
+        }
+        .windowResizability(.contentSize)
+        .defaultSize(width: 900, height: 560)
         .windowStyle(.hiddenTitleBar)
         .defaultPosition(.center)
